+++ web/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Hourglass - WEB APP - CHANGELOG
 
+## Version 1.4.0 - More Sizes and Time - 23/10/2024
+
+- Provide more database size information
+- display the timestamp for historical statistics
+
 ## Version 1.3.0 - Server Name - 18/10/2024
 
 - Now display the server name instead of the server_id